add DHCP server
This commit adds a simple DHCP server when starting the web recovery and
changes the webserver's IP address to 10.123.123.1.
The IP addresses offered by the DHCP server are computed from the client's
MAC addresses in the same fashion as dnsmasq's so that booting OpenWrt with
its default configuration does not cause any problems as the client keeps
its IP address.